Fix releasing lock during the creation of `VmMapping`
