+++ kernel/src/vm/vmar/dyn_cap.rs
@@ -4,7 +4,7 @@ use core::ops::Range;
 
 use aster_rights::Rights;
 
-use super::{vm_mapping::VmarMapOptions, VmPerms, Vmar, VmarRightsOp, Vmar_};
+use super::{VmPerms, Vmar, VmarMapOptions, VmarRightsOp, Vmar_};
 use crate::{
     prelude::*, thread::exception::PageFaultInfo, vm::page_fault_handler::PageFaultHandler,
 };
